Fixed races between concurrent PushBacks
I introduced an atomic_uint32 which holds a count of number of threads
trying to add a datum to the array.

This is the race condition that this solved:
Imagine two threads doing PushBack at the same time. One thread checks
and sees that there is room to place its data in the array and
increments the array size counter. The other thread gets that
incremented size and sees there is no more space; a resize is needed.
There is a case that between the time that the first thread gets the
pointer in the array to place its data, the other thread deletes that
array before the store happens.

The placing counter solves this by making sure all stores are complete
before the resize starts