Fixed a possible race condition in PopFront
Originally, PopFront was capped at size, if the head index reached size
then there were no more elements to pop. The problem with this is that
size is incremented in PushBack before the data is placed in the array.
This leads to a possible case where PopFront reads from an array
location where data has not yet been placed.

To fix this I added another atomic uint called tail. Tail is like size
but is incremented after the data is placed into the array. This makes
sure that PopFront will only read from valid locations in the array.

I also now see that resetting the head and size (now tail too) in
PopFront is generally not a good idea. It might be possible, but it is
not as elegant of a solution (I think) as follows...

When PushBack does a resize we can first check and see how much of the
array has already been popped off. We do this by looking at the value of
head. Since head is incremented every time PopFront is called, all
values from index zero up to index head are values that will not be used
again. If these unused values make up a certain percentage (currently
undecided), say 50% or more, we can say that we don't need to resize the
array. Instead we can simply set head to zero and size and tail to head
- size.

It might not sound simply considering the amount of words it took to
explain it, but I think it will work.