Implemented new way of resizing and resetting in PushBack
I stripped out the resetting from PopFront and placed it in PushBack.
While doing so I noticed some mistakes I made with resetting and
resizing and corrected them. The biggest mistake was when resizing the
array. I did not set the head back to index zero and then also copied
all of the already popped elements into the new, larger array.

Next thing to do is make something to delete/destruct already popped
elements in the array before they are over written. After that I need to
make PopFront not have data races with PushBack and other PopFronts.